requirements.txt
Django
pandas
djangorestframework